# compiler: handle missing LLVM math intrinsics on wasm targets