This enables WebAssembly targets (unknown, wasip2) to use package math without linking libc.

The missing intrinsics are:

```
llvm.exp.f64
llvm.exp2.f64
llvm.log.f64
```

This is required for and cherry-picked from the WASI Preview 2 fork we’re working on with @dgryski.