@ydnar 2024-03-02
It could be interesting to add //export annotations to certain tests in testdata so they can be run under wasm32-unknown.