@aykevl 2024-03-02
While one possible solution, I'd actually much prefer to use a very stripped down libc anyway (for consistency). This libc might only implement these math and perhaps memory builtins (memset, memcpy, memmove).

LLVM is built in a way that require a libc. If you go without, it might work for a bit but eventually you'll run into issues.

If you really don't want a libc, can you explain why?