@ydnar 2024-03-02
@dgryski and I are working on WASI Preview 2 support for TinyGo (GOOS=wasip2). To our knowledge, this is the first native language target of Preview 2 that does not require an adapter shim.

wasi-libc does not yet support WASI Preview 2, and there is not yet a clear timeline when it will.

Second, there are reasons we may want to avoid using it for Go due to how the Component Model interacts with the runtime and GC. 

Given that big Go has non-asm implementations of all math functions, and LLVM has intrinsics for most, it seems reasonable to avoid linking libc (or even libm) for two symbols.

Happy to discuss this in detail with you and @dgryski.