@aykevl 2024-03-03
> wasi-libc does not yet support WASI Preview 2, and there is not yet a clear timeline when it will.

That wouldn't matter for math builtins. They don't need any WASI support (v2 or otherwise) - they're just pure functions.

> Second, there are reasons we may want to avoid using it for Go due to how the Component Model interacts with the runtime and GC.
> 
> Given that big Go has non-asm implementations of all math functions, and LLVM has intrinsics for most, it seems reasonable to avoid linking libc (or even libm) for two symbols.

You'll also lose out on many optimization opportunities. Many math builtins are supported directly in wasm and the only way to use those is by using LLVM builtins. Moreover, LLVM knows these builtins and for example converts the equivalent of `float32(sqrt64(n))` to `sqrt32(float32(n))` (pseudocode).

To be clear: I'm only talking about the math and memory builtins. I understand that most other libc features will need to be reimplemented in some other way. There's no reason to put all these functions together in a single library called "libc".

I can make a PR for this if you're interested, to show what I mean.