@ydnar 2024-03-03
Huh. I'm confused then. This PR doesn't inhibit LLVM intrinsics, only skipping those that aren't present in Wasm.

Maybe it's worth benchmarking the libc exp/exp2 against Go's under wasm?