@dgryski 2024-03-03
We could also add implementations that call into native wasm instructions to our mini wasip2 libc.