@aykevl 2024-03-03
> We could also add implementations that call into native wasm instructions to our mini wasip2 libc.

That's not necessary. LLVM will do that for you. We need separate implementions for the ones that aren't supported by wasm, such as `log` and `exp`.

(I'm working on a math-only version of wasi-libc but of course I run into weird toolchain issues I need to sort out).