@aykevl 2024-03-03
> Edit: as for memory intrinsics, does LLVM lean on libc for those on wasm? It works in our wasip2 branch with no memcpy impl or libc.

Yes. But if you have enabled bulk memory operations, LLVM will use those instead (because they'll be faster than what you can implement in a libc). So it's only needed for environments where bulk memory operations aren't available.