@ydnar 2024-03-04
Benchmarks:

**Before vs after (this PR):**

```
benchmark             old ns/op     new ns/op     delta
BenchmarkMathAbs      3.52          3.91          +11.26%
BenchmarkMathCeil     3.52          3.97          +12.83%
BenchmarkMathExp      3.75          13.0          +247.56%
BenchmarkMathExp2     3.53          14.5          +310.03%
BenchmarkMathLog      3.75          4.50          +19.98%
```

**Before vs after (with pure Go `exp`, `exp2` and `log`:**

```
benchmark             old ns/op     new ns/op     delta
BenchmarkMathAbs      3.52          3.53          +0.43%
BenchmarkMathCeil     3.52          3.62          +2.64%
BenchmarkMathExp      3.75          13.5          +259.03%
BenchmarkMathExp2     3.53          15.1          +326.75%
BenchmarkMathLog      3.75          4.52          +20.46%
```