@ydnar 2024-03-04
Will close this in favor of #4176 when it’s ready.