@dgryski 2024-03-28
https://github.com/tinygo-org/tinygo/pull/4176 has been merged.  This can be closed.